requirements.txt
langchain
langchain-groq
huggingface_hub
streamlit
python-dotenv